+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [4.3.1](https://github.com/netlify/primitives/compare/dev-utils-v4.3.0...dev-utils-v4.3.1) (2025-11-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @netlify/types bumped from 2.1.0 to 2.2.0
+
 ## [4.3.0](https://github.com/netlify/primitives/compare/dev-utils-v4.2.0...dev-utils-v4.3.0) (2025-10-10)